Move UGRID benchmark data generation.
